Almost done with creating input files. In order for this to be possible, the original logical flow of the Boris push algorithms had to be modified. Instead of taking a hard coded array for the input, now we need to dynamically read a file and send it along the functions.
Furthermore, logic for graphing has to be adjusted to account for multiple particles.

One area of concern is how the bulk of the iterative logic happens in for loops; when it comes to test multiple trajectories, it will be problematic
to have to wait for each particle to finish looping. It also bottlenecks multithreading if different threads have to wait for the loop to complete anyway.

There has to be a better way.

TODO:
- Fix formatting of output file
- Create a way for output file names to be dynamically created
- Create methods for reading said files

- Address computational efficiency concerns