reorganization - magpy4c1 had a bad structure of running initialization unexpectedly when being imported by main. - I also needed to ensure that the global variables the GUI ultimately feeds it were copied, but not redundantly initiated.
This is all alluding to the fact that custom input files were not working correctly due to these errors.

To solve these issues, I moved the output file constructors to GuiHelpers.py. I then added an initializer argument to magpy4c1's processpool.
I also shifted the logic so that main is coupled with the GUI, and thus supplies data to magpy4c1 as a middleman. Just reducing coupling.