overhaul of plotting window.
when selecting a dataset, the program will automatically populate the trajectory graph.

Furthermore, there are now two graphs to the side that provide preset graphs available from a dropdown menu.

This was made possible by rewriting the plotting window class with an object oriented approach; graphical styles (such as titles, axis labels, so on), graphing functions, and specifics are tucked away in their unique functions and dictionaries.

The actual graph widget objects contain an update function that takes in a callable as a parameter. This allows for the class to call different graphing logics without needing to specify the exact function being used.

We instead specify the graphing function used when we invoke the update function.

The basic functionality is there and works. From here, I can work on enhancing visuals with extra qol (toolbars, legends, reducing number of points on the trajectory graph to increase performance, etc.)