Inputs/outputs now defauls to '~/Documents/Boris_usr'.
I also added more to the tempfile dictionary, as it now also tracks the full path to the particle condition file as well as the coil file.

Because the creation of default input files (when there are no files in the respective dir) now uses the mirror preset instead of being a single default value coil called 'Default_Dan', I also had to change some internal workings to account for the fact that the file creation and saving would take place before the table's __init__() executes fully.