handling added for documents folders that exist externally. Also creates a new default config everytime on startup.
Need to find a way to explicitly set the python path for this project to include /Scripts, as right now it depends on ide specific worlplace settings. This introduces an extra headache layer for working with paths and import statements, as it has to be manually set up every time you clone this repo.

This is very annoying because VSCode's environment PYTHONPATH variable refuses to work with multiple paths sometimes. I want to add both the project root and Scripts/, but it has a hard time running with those in mind. The IDE totally finds it though. Very annoying.